Fix bug
Разрешила merge конфликт
