--- src/main/java/dictionary/configuration/DictionaryFactory.java
@@ -1,22 +1,0 @@
-package dictionary.configuration;
-
-import dictionary.storage.DictionaryStorage;
-import dictionary.storage.FileStorage;
-import dictionary.storage.MapStorage;
-
-/**
- * Создаёт стораджи
- */
-public class DictionaryFactory {
-
-    private final static String MAP = "map";
-
-    public static DictionaryStorage createDictionaryStorage(String arg) {
-        if (arg.equals(MAP)) {
-            return new MapStorage();
-        } else {
-            return new FileStorage();
-        }
-    }
-
-}
