+++ src/main/java/dictionarySpring/exception/TypeNotFoundException.java
@@ -1,4 +1,4 @@
-package dictionary.exception;
+package dictionarySpring.exception;
 
 public class TypeNotFoundException extends Exception{
     public TypeNotFoundException(String message){
