+++ src/main/java/dictionarySpring/model/DictionaryLine.java
@@ -1,9 +1,14 @@
-package dictionary.model;
+package dictionarySpring.model;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+@Component
 public class DictionaryLine {
 
     private final String key;
     private final String value;
+
     public DictionaryLine(String key, String value) {
         this.key = key;
         this.value = value;
