--- src/main/java/dictionarySpring/service/DictionaryException.java
@@ -1,8 +1,0 @@
-package dictionarySpring.service;
-
-public class DictionaryException extends Exception{
-    public DictionaryException(String message){
-        super(message);
-    }
-
-}
