+++ src/main/java/dictionarySpring/service/DictionaryService.java
@@ -1,42 +1,93 @@
 package dictionarySpring.service;
 
 import dictionarySpring.configuration.DictionaryType;
+import dictionarySpring.model.DictionaryLine;
 import dictionarySpring.storage.*;
-import dictionarySpring.validator.ValidInterface;
+import dictionarySpring.validator.Validator;
+import dictionarySpring.view.Formation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
+import java.util.Optional;
 
+import static dictionarySpring.view.Console.NO_EXIST_KEY;
+
+/**
+ * Класс отвечает за разделение слоя хранения и слоя представления
+ */
 @Component
 public class DictionaryService {
 
-    private final ValidInterface validInterface;
+    private final Validator validator;
     private final DictionaryStorage dictionaryStorage;
 
     @Autowired
-    public DictionaryService(ValidInterface validInterface, DictionaryStorage dictionaryStorage) {
-        this.validInterface=validInterface;
+    private Formation formation;
+
+    @Autowired
+    public DictionaryService(Validator validator, DictionaryStorage dictionaryStorage) {
+        this.validator = validator;
         this.dictionaryStorage = dictionaryStorage;
     }
 
-    public String addService(String key, String value, DictionaryType selectedDictionary) {
-        if (validInterface.isValidPair(key, value, selectedDictionary)) {
-            return dictionaryStorage.add(key, value, selectedDictionary);
+    /**
+     * Метод отвечает за валидацию введенный данных, и при успехе обращается к методу добавления относительно выбранного способа хранения
+     * и выбранного языка
+     * и возвращает true, при ошибке валидации возвращает false
+     *
+     * @param key                ключ, введенный пользователем
+     * @param value,             значение введенное пользователем
+     * @param selectedDictionary выбранный язык словаря
+     * @return логическое значение
+     */
+    public boolean addService(String key, String value, DictionaryType selectedDictionary) {
+        if (validator.isValidPair(key, value, selectedDictionary)) {
+            dictionaryStorage.addTo(key, value, selectedDictionary);
+            return true;
         } else {
-            return "Error";
+            return false;
         }
     }
 
+    /**
+     * Метод отвечает за обращение к методу чтения данных относительно способа хранения и выбранного языка
+     *
+     * @param selectedDictionary выбранный язык словаря
+     * @return строки из хранилища
+     */
     public List<String> readService(DictionaryType selectedDictionary) {
-        return dictionaryStorage.read(selectedDictionary);
+        return formation.castToString(dictionaryStorage.read(selectedDictionary));
     }
 
-    public void removeService(String key, DictionaryType selectedDictionary) {
-        dictionaryStorage.remove(key, selectedDictionary);
+    /**
+     * Метод отвечает за оборащение к методу удаление данных относительно способа хранения и выбранного языка
+     *
+     * @param key                ключ, введенный пользователем
+     * @param selectedDictionary выбранный язык словаря
+     * @return логическое значение
+     */
+    public boolean removeService(String key, DictionaryType selectedDictionary) {
+        return dictionaryStorage.remove(key, selectedDictionary);
     }
 
-    public String searchService(String key, DictionaryType selectedDictionary) {
-        return dictionaryStorage.search(key, selectedDictionary);
+
+    /**
+     * Метод отвечает за обращение к методу поиска значения по ключу и вывод строки относительно способа хранения выбранного языка
+     *
+     * @param key                ключ, введенный пользователем
+     * @param selectedDictionary выбранный язык словаря
+     * @return объект типа DictionaryLine
+     */
+    public String searchService(String key, DictionaryType selectedDictionary)
+    {
+        final Optional<DictionaryLine> optionalReturn = Optional.ofNullable(dictionaryStorage.search(key, selectedDictionary));
+        if (optionalReturn.isEmpty()) {
+            return NO_EXIST_KEY;
+        }
+        else {
+            return formation.castToString(dictionaryStorage.search(key, selectedDictionary));
+        }
     }
-    }
+}
+
