+++ src/main/java/dictionarySpring/storage/DictionaryStorage.java
@@ -1,12 +1,21 @@
 package dictionarySpring.storage;
 
 import dictionarySpring.configuration.DictionaryType;
+import dictionarySpring.model.DictionaryLine;
 
 import java.util.List;
 
+/**
+ * Интерфейс предоставляющий метододы для работы с данными словаря
+ */
 public interface DictionaryStorage {
-    List<String> read(DictionaryType selectedDictionary);
-    String add(String key, String value, DictionaryType selectedDictionary);
-    void remove(String key, DictionaryType selectedDictionary) ;
-    String search(String key, DictionaryType selectedDictionary);
+
+    List<DictionaryLine> read(DictionaryType selectedDictionary);
+
+    boolean addTo(String key, String value, DictionaryType selectedDictionary);
+
+    boolean remove(String key, DictionaryType selectedDictionary);
+
+    DictionaryLine search(String key, DictionaryType selectedDictionary);
+
 }
