+++ src/main/java/dictionarySpring/storage/FileStorage.java
@@ -1,101 +1,140 @@
 package dictionarySpring.storage;
 
 import dictionarySpring.configuration.DictionaryType;
-import org.springframework.core.io.ClassPathResource;
+import dictionarySpring.model.DictionaryLine;
+import dictionarySpring.service.DictionaryLineCodec;
 
 import java.io.*;
-import java.util.ArrayList;
+import java.util.LinkedList;
 import java.util.List;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 
-
+/**
+ * Класс отвечающий за хранение словаря в файловой системе
+ */
 public class FileStorage implements DictionaryStorage {
 
+    private DictionaryLineCodec dictionaryLineCodec = new DictionaryLineCodec();
 
-    private static final String ADD_KEY = "added";
-
-    public void fileClear(String path) {
+    private void fileClear(String path, boolean isClear) {
         try {
-            new FileWriter(path, false).close();
-        } catch (IOException ioException) {
-            ioException.printStackTrace();
-        }
-    }
- private  File getDictionaryFile(String path) throws IOException {
-     return new ClassPathResource(path).getFile();
- }
-
-    public String write(String key, String value, String path) {
-
-        BufferedWriter bufferedWriter = null;
-        try {
-            FileWriter fileWriter = new FileWriter(getDictionaryFile(path), UTF_8, true);
-            bufferedWriter = new BufferedWriter(fileWriter);
-            bufferedWriter.write(key + DictionaryType.getSymbol() + value + "\n");
-            bufferedWriter.flush();
+            new FileWriter(path, isClear).close();
         } catch (IOException e) {
             e.printStackTrace();
-        } finally {
-            try {
-                bufferedWriter.close();
-            } catch (Exception e) {
-            }
         }
-        return ADD_KEY;
     }
 
-    public List<String> operationRead(String path) {
+    /**
+     * Метод, который отвечает за запись в файл
+     *
+     * @param key   - ключ
+     * @param value - значение
+     * @param path  - принимает путь
+     */
+    private void write(String key, String value, String path, boolean isWrite) throws IOException {
 
-        List<String> results = new ArrayList<String>();
-        try {
+        try (BufferedWriter writer = new BufferedWriter(new FileWriter(path, UTF_8, isWrite))) {
+            DictionaryLine dictionaryLine = new DictionaryLine(key, value);
+            writer.write(dictionaryLineCodec.decode(dictionaryLine) + "\n");
+        }
+    }
 
-            BufferedReader reader =  new BufferedReader(new InputStreamReader(new FileInputStream(getDictionaryFile(path)), UTF_8));
-            String line = reader.readLine();
-            while (line != null) {
-                results.add(line);
-                line = reader.readLine();
+    /**
+     * Метод, который отвечает за доступ к чтению из файла
+     *
+     * @param path - принимает путь
+     * @return - возвращает список <ключ,значение>
+     */
+    private List<DictionaryLine> operationRead(String path) {
+
+        List<DictionaryLine> results = new LinkedList<>();
+        try (BufferedReader reader = new BufferedReader(new FileReader(path))) {
+            String line;
+            while ((line = reader.readLine()) != null) {
+                results.add(dictionaryLineCodec.encode(line));
             }
-        } catch (FileNotFoundException e) {
-            e.printStackTrace();
         } catch (IOException e) {
             e.printStackTrace();
         }
         return results;
     }
 
+    /**
+     * Метод, который отвечает за чтение данных из файла
+     *
+     * @param selectedDictionary - принимает вид языка с которым работает
+     * @return mapRead - возвращает список пар <ключ, значение>
+     */
     @Override
-    public List<String> read(DictionaryType selectedDictionary) {
+    public List<DictionaryLine> read(DictionaryType selectedDictionary) {
         return operationRead(selectedDictionary.getDictionaryPath());
     }
 
+    /**
+     * Метод, который отвечает за добавление данных в файл
+     *
+     * @param key                - ключ
+     * @param value              - значение
+     * @param selectedDictionary - принимает вид языка с которым работает
+     * @return логическое значение
+     */
     @Override
-    public String add(String key, String value, DictionaryType selectedDictionary) {
-        return write(key, value, selectedDictionary.getDictionaryPath());
+    public boolean addTo(String key, String value, DictionaryType selectedDictionary) {
+        try {
+            write(key, value, selectedDictionary.getDictionaryPath(), true);
+        } catch (IOException e) {
+            e.printStackTrace();
+            return false;
+        }
+        return true;
     }
 
+    /***
+     * Метод, который отвечает за удаление данных из файла
+     * @param key - ключ
+     * @param selectedDictionary - принимает вид языка с которым работает
+     * @return логическое значение
+     */
     @Override
-    public void remove(String key, DictionaryType selectedDictionary) {
-
-        List<String> readLines = operationRead(selectedDictionary.getDictionaryPath());
-        for (int i = 0; i < readLines.size(); i++) {
-            if (key.equals(readLines.get(i).split(DictionaryType.getSymbol())[0])) {
-                readLines.remove(i);
+    public boolean remove(String key, DictionaryType selectedDictionary) {
+        boolean isRemoved = false;
+        List<DictionaryLine> readLines = operationRead(selectedDictionary.getDictionaryPath());
+        for (DictionaryLine dictionaryLine : readLines) {
+            if (dictionaryLine.getKey().equals(key)) {
+                isRemoved = readLines.remove(dictionaryLine);
                 break;
             }
         }
-        fileClear(selectedDictionary.getDictionaryPath());
-        for (String readLine : readLines) {
-            String[] keyAndValue = readLine.split(DictionaryType.getSymbol());
-            write(keyAndValue[0], keyAndValue[1], selectedDictionary.getDictionaryPath());
+        if (!isRemoved) {
+            return false;
         }
+        fileClear(selectedDictionary.getDictionaryPath(), false);
+        for (DictionaryLine readLine : readLines) {
+            try {
+                write(readLine.getKey(), readLine.getValue(), selectedDictionary.getDictionaryPath(), true);
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
+        return true;
     }
 
-    public String search(String key, DictionaryType selectedDictionary) {
-        List<String> searchLines = operationRead(selectedDictionary.getDictionaryPath());
-        for (int i = 0; i < searchLines.size(); i++) {
-            if (key.equals(searchLines.get(i).split(DictionaryType.getSymbol())[0])) {
-                return searchLines.get(i);
+    /**
+     * Метод, который отвечает за поиск данных в файле
+     *
+     * @param key                - ключ
+     * @param selectedDictionary - принимает вид языка с которым работает
+     * @return mapRead - возвращает список пар <ключ, значение> и если ключ не найден
+     * возвращается null
+     */
+
+    @Override
+    public DictionaryLine search(String key, DictionaryType selectedDictionary){
+        List<DictionaryLine> searchLines = operationRead(selectedDictionary.getDictionaryPath());
+        for (DictionaryLine searchLine : searchLines) {
+            if (key.equals(searchLine.getKey())) {
+                return searchLine;
             }
         }
         return null;
