--- src/main/java/dictionarySpring/validator/ValidInterface.java
@@ -1,7 +1,0 @@
-package dictionarySpring.validator;
-
-import dictionarySpring.configuration.DictionaryType;
-
-public interface ValidInterface {
-    boolean isValidPair(String key, String value, DictionaryType dictionaryType);
-}
